trip-planner/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==1.1.0
openai==0.27.8
langgraph==0.0.23
requests==2.26.0
python-dateutil==2.8.1
pydantic==1.10.16
